pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "numpy (>=2.3.5,<3.0.0)",
    "datasets (>=4.4.1,<5.0.0)",
    "trl (>=0.25.1,<0.26.0)",
    "peft (>=0.18.0,<0.19.0)",
    "sacrebleu (>=2.5.1,<3.0.0)",
    "rouge (>=1.0.1,<2.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "bert-score (>=0.3.13,<0.4.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "lmstudio (>=1.5.0,<2.0.0)",
    "llama-cpp-python (>=0.3.16,<0.4.0)"
]

[tool.poetry]
packages = [{include = "final_project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
